MCPServer/pyproject.toml
[project]
name = "unreal-agent-mcp"
version = "0.1.0"
description = "MCP server for UnrealAgent - AI control interface for Unreal Editor"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "Pillow>=10.0.0",
]

[project.scripts]
unreal-agent-mcp = "unreal_agent_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unreal_agent_mcp"]
